S08P12A104-220 fix : search error 수정

+++ back/src/main/java/com/justudy/backend/study/service/StudyResumeService.java
@@ -42,10 +42,12 @@ public class StudyResumeService {
     }
 
     @Transactional
-    public void deleteStudyResume(Long id) {
+    public void deleteStudyResume(Long id, Long loginSequence) {
         StudyResumeEntity studyResumeEntity = studyResumeRepository.findById(id).orElseThrow(InvalidRequest::new);
         StudyEntity studyEntity = studyRepository.findById(studyResumeEntity.getStudy().getSequence())
                 .orElseThrow(StudyNotFound::new);
+        if (studyResumeEntity.getMember().getSequence() != loginSequence)
+            throw new InvalidRequest();
         studyEntity.removeStudyResume(studyResumeEntity);
         studyResumeRepository.deleteById(id);
     }
@@ -57,9 +59,11 @@ public class StudyResumeService {
     }
 
 
-    public List<StudyResumeResponse> readAllStudyResumeByStudy(Long id) {
-        studyRepository.findById(id)
+    public List<StudyResumeResponse> readAllStudyResumeByStudy(Long id, Long loginSequence) {
+        StudyEntity studyEntity = studyRepository.findById(id)
                 .orElseThrow(StudyNotFound::new);
+        if (loginSequence != studyEntity.getLeaderSeq()) throw new InvalidRequest();
+
         return studyResumeRepository.readAllStudyResumeByStudy(id)
                 .stream()
                 .map(StudyResumeResponse::makeBuilder)
@@ -72,6 +76,10 @@ public class StudyResumeService {
                 .map(StudyResumeEntity::getStudy)
                 .map(StudyResponse::makeBuilder)
                 .collect(Collectors.toList());
+    }
+
+    public void deleteById(Long id) {
+        studyResumeRepository.deleteById(id);
     }
 }
